requirements.txt
absl-py==1.2.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cloudpickle==2.1.0
cycler==0.11.0
decorator==5.1.1
dm-acme==0.4.0
dm-env==1.5
dm-launchpad==0.5.0
dm-reverb==0.7.0
dm-sonnet==2.0.0
dm-tree==0.1.7
etils==0.6.0
executing==0.8.3
flatbuffers==2.0
fonttools==4.34.4
gast==0.5.3
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gym==0.25.0
gym-notices==0.0.7
h5py==3.7.0
idna==3.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipython==8.4.0
jax==0.2.22
jaxlib @ file:///home/mrl/Robolympics/jaxlib-0.1.69%2Bcuda111-cp38-none-manylinux2010_x86_64.whl
jedi==0.18.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.1
Markdown==3.4.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mock==4.0.3
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkg_resources==0.0.0
portpicker==1.5.2
prompt-toolkit==3.0.30
protobuf==3.20.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.8.1
six==1.16.0
stack-data==0.3.0
tabulate==0.8.10
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-probability==0.15.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
traitlets==5.3.0
trfl==1.2.0
typing_extensions==4.3.0
urllib3==1.26.10
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
